Mode1が使用されたときに追加されるパラメータを読み取りNoteに追加するようにした また、NoteをGenerate.csに移動した
